Fix type conversion issues for TF 2.4.4

+++ openvino_tensorflow/encapsulate_clusters.cc
@@ -470,7 +470,7 @@ Status Encapsulator::RewritePass(
             .Attr("Targuments", input_types)
             .Attr("Tresults", cluster_output_dt_map[cluster_idx])
             .Attr("ngraph_graph_id", graph_id)
-            .Attr("cluster_cost", cluster_cost_map_in_ms[cluster_idx])
+            .Attr("cluster_cost", (int)cluster_cost_map_in_ms[cluster_idx])
             .Device(device_name_map[cluster_idx])
             .Input(inputs);
     if (!device_config.empty()) {
